Document thread-unsafe iterator access to resource maps
By using a proxy type, everywhere resource maps are iterated are clearly
marked as not being thread safe.  In most cases, only destruction and
capture/replay iterate over these maps, which means thread safety is not
an issue (or is externally enforced).

The only case where iterators are used in the presence of other contexts
is with ANGLE_request_extension, which is changed to explicitly require
the application to ensure thread safety.  In practice, the user is
Chrome which already guarantees this.

Bug: angleproject:8667
Change-Id: I7af13c6433b6955d9c36f9088b3aa4c065e1cfc1
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5526428
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>