Vulkan: Fix AtomicRefCounted threading issue
The `AtomicRefCounted::getAndReleaseRef()` is used to decrement the
reference counter and to destroy the object if there are no more uses.
Decrement operation must use acquire-release memory ordering, in order
to guarantee that all accesses from other threads happened before last
thread checks that counter is "1" and destroys the object.

Added comments to `releaseRef()` and `isReferenced()`, that these
methods do not provide any synchronization, and can not be used to
achieve same affect as `getAndReleaseRef()`. Relaxed ordering is kept
for performance reasons.

Bug: angleproject:8601
Change-Id: I8e5654416383e9490c23bc02fda9664216a81076
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5573394
Reviewed-by: Charlie Lao <cclao@google.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Igor Nazarov <i.nazarov@samsung.com>