Translator: Fix internal inconsistency with output init
When creating "zero" nodes, this transformation directly takes the
TType's from the variables being initialized.  This means that
qualifiers such as invariant, as well as any associations with interface
blocks are retained in these constructors, which is not valid.

This can trip up the SPIR-V generator which does not expect constructors
to be qualified as such (i.e. assumes the TypeSpec to be the default).

Bug: angleproject:342866364
Bug: angleproject:342866374
Bug: angleproject:343011429
Change-Id: If5c38e18ce977bd0eb1a48e41c59f6199eb7d2e5
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5570155
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>