CL: Fix entrypoint validation for ICDs
For some OpenCL ICD implementations,
the clGetPlatformIDs API is left as nullptr
in the ICD dispatch table. This is because
the OpenCL ICD Loader instead uses the
clIcdGetPlatformIDsKHR call instead.

Updating the passthrough code to only
require checking the clGetPlatformIDs API
if the platform is not an ICD.

Bug: angleproject:42266872
Change-Id: Ic723adbf3ad0f6c79389ef629dd24ae8a82ca9ba
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5789651
Commit-Queue: Austin Annestrand <a.annestrand@samsung.com>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>