Vulkan: Prefer fallback to renderable texture than CPU copy
When a texture copy is done, the transfer path is preferred.  If
transfer is not possible, ANGLE falls back to a draw-based copy.
However, if the format is not renderable, ANGLE falls back to CPU copy.

Instead, this change makes ANGLE try to fall back to a renderable format
before giving up on the draw-based copy.

Bug: b/351872449
Change-Id: I3d09dd88cc0b7feb5e92e58c4f19380973eb5e94
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5775592
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>
Reviewed-by: Charlie Lao <cclao@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>