CL/Vulkan: Add parent dependencies and image layout transitions
In the case of memory objects derived from parent object, the
dependencies need to be set for parent objects as well. Setting up those
dependencies.

In the case of images, there might be a layout transition needed when
using them. Adding an appropriate barrier command to do layout
transitions.

Also, refactored all the common code of setting up dependencies in a
utility function.

Bug: angleproject:380188569
Change-Id: Ia46f8a7fe647bf38c86f1ea28f99d44b0416c334
Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6005390
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>