Vulkan: Remove unnecessary perfetto trace in BufferVk::mapRange
There is not much usefulness to know there is mapRange call. The more
meaningful place is where we end up waiting for GPU to finish which we
already have the perfetto trace in SharedFence::wait. The exitsing of
this trace made a lot of noise in trace since app tends to call this a
lot and there are many tiny lines on the trace.

Bug: None
Change-Id: I83fbbba827361d068b521bdfb2bc187865d11012
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5841067
Commit-Queue: Charlie Lao <cclao@google.com>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>