Fix ignoring blit bits when attachments are missing
GLES requires that the bits provided to glBlitFramebuffer are silently
ignored if either the read or draw framebuffer is missing the
corresponding attachment.

In WebGL, this is forbidden.  ANGLE was forbidding this because of
WebGL, but a recent CTS test revealed that we shouldn't be doing that
when !WebGL.

Bug: angleproject:370917919
Change-Id: I202fbf665d8a102f0aea72709b80dfdabeb71a2e
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5912614
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>