Fix for float constant precision in the GLSL backend.
Increase the precision of floating point values written out
via std::ostringstream. 8 digits is not sufficient to represent
all floating point values.

Note: the reason the locale test was modified is because it was using
a value of 1.9, which has no exact fp32 representation. Increasing
the precision causes it to print as 1.8999998 instead of 1.9, failing
the test. I've adjusted the value to 1.5, since this does have an
exact fp32 representation. (However, note that I couldn't get the test
to fail when I removed the locale setting, with either 1.9 or 1.5.
Perhaps the locale is being handled at a different level.)

Bug: angleproject:374013421
Change-Id: Icb79eb9acd562c83d079f2cc2cdba253220e581e
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5938473
Commit-Queue: Stephen White <senorblanco@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>