Vulkan: Fix `vec4(...).zxwy[index]`
The code was actually correct, but included an unnecessary assertion.
The code incorrectly assumed `.zxwy[index]` is pre-processed with a
lookup table and turned into `[index']`, but it wasn't if the LHS is a
constructor.

The change just removes the assertion because it was correctly handling
that already.

Bug: angleproject:40096715
Change-Id: Ib1365f95a255ce3654831af55429361d35c026c4
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5949527
Reviewed-by: Yuxin Hu <yuxinhu@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>