Improve CanSupportAEP Error Reporting
This change improves the error reporting for the `CanSupportAEP`
function.  Previously, if the function returned false, it was difficult
to determine which requirement was not met. This change adds logging
that will list all unmet requirements, making it easier to debug AEP
support issues.

Test: the default presubmit
Bug: b/268108262
Change-Id: I48eb6b26358ca4991d760523f44568d97e8c35a6
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/5972497
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Solti Ho <solti@google.com>
Reviewed-by: Yuxin Hu <yuxinhu@google.com>