Tests: linux-trace logs return code with OOM note
In case on the bug the replay was terminated with an empty stdout
and signal handler wouldn't trigger, making it difficult to
diagnose. This happened due to OOM killer (system low on memory) sending
SIGKILL to the process, which can't be handled and results in
termination with exit code -9. Add a note to the logs when this happens.

Bug: angleproject:380296979
Change-Id: Id4db4b12c680701511430d90eee847baded03bcb
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6042183
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>