Capture: add a sanity check to InitializeReplay4 args
In the case on the bug we were getting
maxProgramPipeline = 1684105299
due to an unitialized memory read, which then led to Replay OOM crash
as it allocates an array of this size. This is difficult to diagnose in
Replay as OOM leads to SIGKILL which cannot be trapped so we can't have
a backtrace. Adding a sanity check during capture.

Fix the other tests where we hit this
(https://anglebug.com/380296979#comment6)
One of them visible on red tests on a previous patchset in this CL:
https://ci.chromium.org/ui/p/angle/builders/try/linux-trace/9871/overview
https://ci.chromium.org/ui/p/angle/builders/try/win-trace/10795/overview

Bug: angleproject:380296979
Change-Id: I2879c1947742a9751a122545c8ba8da23ab243cb
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6042185
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Commit-Queue: Roman Lavrov <romanl@google.com>